clean up to match ng2 stylguide and cli, gulp task work
I added the gulp task that adds all the functionality we need to build
out the app easily. I can convert it back to TS once I'm done. I need to
tweak it to drop in angular universal so the node server is turned off
right now and it is going on hash routes.